+++ CodeStarterCamp_Week4/Routine.swift
@@ -27,19 +27,24 @@ struct Routine {
 func doRoutine(routine: Routine, bodyCondition: BodyCondition) throws {
     
     let countArray = ["첫", "두", "세", "네", "다섯","여섯", "일곱"]
-        
-        print("루틴을 몇 번 반복할까요?")
-        let inputString = readLine()!
-        
     
-    guard inputString.allSatisfy({$0.isNumber}) else {
+    print("루틴을 몇 번 반복할까요?")
+    guard let inputString = readLine() else {
+        return
+    }
+    
+    
+    guard inputString.allSatisfy({ $0.isNumber }) else {
         throw RoutineError.inputError
     }
     
-    let inputNumber = Int(inputString)!
+    guard let inputNumber = Int(inputString) else{
+        print("입력값이 잘못되었습니다.")
+        return
+    }
     print("-------------")
-        
-    for number in 0..<inputNumber{
+    
+    for number in 0..<inputNumber {
         print("\(countArray[number]) 번째 (을)를 시작합니다.")
         routine.intermediate(bodyCondition: bodyCondition)
         
@@ -52,20 +57,5 @@ func doRoutine(routine: Routine, bodyCondition: BodyCondition) throws {
     
 }
 
-enum RoutineError: Error {
-    case overTirednessError
-    case inputError
-}
 
-func doRoutineWithCatchingError(routine: Routine, bodyCondition: BodyCondition) {
-    do {
-        try doRoutine(routine: routine, bodyCondition: bodyCondition)
-    } catch RoutineError.overTirednessError {
-        print("피로도가 100 이상입니다. 루틴을 중단합니다.")
-        bodyCondition.printCondtion()
-    } catch RoutineError.inputError {
-        print("잘못된 입력 형식입니다. 다시 입력해주세요.")
-    } catch {
-        print("default error catch")
-    }
-}
+
