+++ CodeStarterCamp_Week4/main.swift
@@ -1,10 +1,5 @@
 import Foundation
 
-
-//func printResult(activity: Activity) {
-//    activity.action(bodyCondition)
-//    bodyCondition.printCondtion()
-//}
 let routine = Routine()
 var bodyCondition = BodyCondition()
